refactor: split hypervisor/backend.go by operation
Move methods out of the 1000-line monolith into per-operation files
matching SKILL.md's 'one responsibility per file' rule. backend.go
keeps the const block, Backend/BackendConfig types, and the *Spec
hooks; everything else moves to inspect.go, state.go, create.go,
clone.go, restore.go, start.go, stop.go, helpers.go (new), plus
appended sections in snapshot.go and gc.go.
